requirements.txt
openai>=1.0.0
python-dotenv
requests
sendgrid
docx2txt
spacy
scikit-learn
numpy
